requirements.txt
alembic
Authlib
blinker
cachelib
certifi
cffi
charset-normalizer
click
colorama
cryptography
dnspython
email-validator
Flask
Flask-Login
Flask-Migrate
Flask-Session
Flask-SQLAlchemy
Flask-WTF>=1.2.1
greenlet
idna
itsdangerous
Jinja2
Mako
MarkupSafe
msgspec
psycopg2-binary
pycparser
python-dotenv
redis
requests
SQLAlchemy
typing_extensions
urllib3
WTForms
playwright
Pillow
Werkzeug>=2.3,<3.0
APScheduler
